Small tweaks after review
